Rework how config is imported
The code would try to insert the right directory early into the search
path and then import the module. This is ugly and fragile: in particular
it doesn't seem to work with pytest, which plays similar games itself.

Let's simplify the whole thing: let's do the import in just one place
in util, and only on demand. Currently __init__.py needs the config when
imported, so this doesn't save much, but if that is refactored, it should
be possible to not load the config until later. Instead of playing with
the search paths, let's just call build the module ourselves.

This allows the tests to pass cleanly under pytest-[23].