Adapt code to be python3 compatible and easier to test (#122)
* 2to3: apply print fixer

* 2to3: get rid of .iteritems()

* 2to3: fix .items() use

* 2to3: apply xrange fixer

* 2to3: only call reload() under python2

* Get rid of unused variable

* Get rid of another unused variable

* run_tests.sh: specify which python to run as environment variable

This way to run the tests is horrible, but let's keep using it
until we have something better.

* 2to3: use relative imports for internal modules

* run_tests.sh: set $MOTE_CONFIG_FOLDER instead of overwriting stuff

* 2to3: variable got renamed in python3

* run_tests: split out the test data creation into a separate script

This makes it easy to just run the tests without the big helper script:
$ rm -rf /tmp/foo && tests/setup-test-dir.sh /tmp/foo
$ TEST_DATA=/tmp/foo MOTE_CONFIG_FOLDER=/tmp/foo/config python2 -m pytest -v
$ TEST_DATA=/tmp/foo MOTE_CONFIG_FOLDER=/tmp/foo/config python3 -m pytest -v

* Rework how config is imported

The code would try to insert the right directory early into the search
path and then import the module. This is ugly and fragile: in particular
it doesn't seem to work with pytest, which plays similar games itself.

Let's simplify the whole thing: let's do the import in just one place
in util, and only on demand. Currently __init__.py needs the config when
imported, so this doesn't save much, but if that is refactored, it should
be possible to not load the config until later. Instead of playing with
the search paths, let's just call build the module ourselves.

This allows the tests to pass cleanly under pytest-[23].

* Remove the obsolete spec file

https://src.fedoraproject.org/rpms/mote/blob/master/f/mote.spec has a
bunch of changes wrt to the copy here. Fedora specifies that dist-git
is the canonical location for the spec file. There isn't much point in
keeping an obsolete version here.

* Default to python3

The code should now be fully version-agnostic, there are only two
places where python version is specified.

* travis: change the Python version from 2.7 to 3.8

* tests: make sure bash is used

It seems we get some strange shell (dash?) in Ubuntu CI.
Let's make sure we get bash so pattern expansion works as expected.