:package: :arrow_up: Pipfile: Upgrade env to Python 3.9
This commit raises Møte's supported Python version to Python 3.9. In the
upcoming Fedora 33 release, Fedora 3.9 becomes the default version of
the Python interpreter. So, this follows Fedora's precedent by using the
latest version of upstream Python.

Since we are still early in development, we can be more flexible about
this for now.

Part of #126, but does not complete it.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>

squash! :package: :arrow_up: Pipfile: Upgrade env to Python 3.9

Update Travis CI build environment to Python 3.9

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>