:wrench: config.py now uses cache file in . directory.
With our current Fedora Infrastructure, we're using a file-based cache storage
we can do the same for our development experience.
while memcached being turned off,
It was trying to write to a file in / root directory resulting in an error (which works fine in the deployed Fedora-infra).

This commit modifies the location of the cache file to create it in the current directory.

It would help developers, who are unable to setup memcached
and would like to use file based cache for development.