Add Sphinx documentation toolchain, CI hooks, and publishing pipeline (#145)
* :memo: :black_nib: ReadTheDocs: Add publishing config and automation

* Add ReadTheDocs config file (not required, but helpful)
* Add `requirements.txt` and script for updating it (required for adding
  table support to a ReadTheDocs virtualenv)

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>

* :new: :tada: :memo: docs: Initial commit

Adds a barebones docs site structure powered by Python Sphinx. Currently
there is not much here. But this will be a base that we can add new
documentation into over time.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>

* :alembic: :new: ci: Add CI job for building and testing Sphinx

This commit expands `.travis.yml` by adding a new job to explicitly run
the docs build script and make sure nothing is broken with our docs
pipeline.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>

* :wrench: Refactor docs dependency management

This commit reorganizes the docs dependencies to sit outside of the
Pipfile and to not use Pipenv. The dependencies can still be installed
inside of a Pipenv shell if necessary.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>

Co-authored-by: Nasir Hussain <nasirhussainm14@gmail.com>